Move polar tests to their own module.
