+++ lib/matplotlib/tests/baseline_images/test_polar/polar_invertedylim.png
